Artifact graph: cascade producer notification
This works around the limitation of synchronous producers not being able
to commit their own versions and those of their outputs because the
datatype registry is not passable to trait object functions. Instead,
producer notifications can return versions that should be committed, and
these are used to cascade commits and thereby production.