Move procmacro-ref'd traits to separate crate
This is necessary to work around traits being referenced in the coupled
procedural macro crate when those macros are called by the crate
providing the trait. There are other workarounds, but they do not work
for the mixture of 2016 and 2018 edition code that is here. The previous
approach of using the trait in a wrapper simple macro does not work
because it fails when the macro is used twice in the same module.