Factor heraclitus-core
Reuse the unnecessary heraclitus-core crate to factor out core datatype
up to and exclusive of artifacts into a separate core crate.

The primary motivation is to simplify testing these core types, because
testing changes will no longer require updating all of the datatypes and
backends implemented by heraclitus. This will also prevent situations
like the current one with StoreRepoBackend, where it is untenable
because extension crates can not implement their own datatypes. The
two-crate arrangement makes this situation immediate and obvious, as now
when the crates cannot be compiled.

One necessary change not yet implemented is the removal of the
MetaController trait. This design is no longer possible because
MetaController depends on Artifact in heraclitus (and impl of it on
backend details of the artifact graph), but core backend traits
depended on MetaController. Instead this can be realized by simply using
the interface system.