Make artifact graphs artifacts
Artifact graphs, being already datatypes, are now artifacts and
datatype::Storage types, rather than existing externally to the rest of
the datatype storage system.

All artifacts and artifact graphs ultimately trace their lineage to the
origin artifact graph. The origin graph is immutable, and exists as an
artifact in itself. All mutable state is in the root artifact graph,
which is an artifact in the origin artifact. When the content of the
root AG change, this changes the root AG artifact and thus the origin
VG, but not the origin AG.

This implementation is not settled and will be improved:
- Ergonomics of creating graphs are tedious.
- The duplication between ArtifactGraphs and ArtifactGraphDescriptions
creates many problems. ArtifactGraphs will most likely lose their
references to the relatively useless Datatype type (since one must have
access to the DatatypesRegistry to make any use of these anyway), making
the ArtifactGraph not have a lifetime and therefore be a valid
StateType.
- Hashes and hash verification around ArtifactGraphs and descriptions
have pending problems.
- The artifact graph controller should be factored.