Remove datatype reference from artifacts
Remove the reference to the datatypes in the registry from artifacts.
This removes the lifetime on the artifact graph, allowing rental
structures to contain both a version graph and the artifact graph the
version graph references, because rental constraints did not satisfy the
"outlives the datatype references" lifetime bounds.

The reference to datatype was not useful because only the datatype's
name was used, which was then looked up in datatypes registry. This is
due to the generic bounds on InterfaceEnum necessary to do anything
useful with datatype models. While a future refactoring fixing many of
the issue with datatypes and datatype registration may reintroduce these
references, generally the coupling between artifacts and datatypes was
less useful (because one may want to do many things with artifacts
without caring about the capabilities of their datatypes), while this
coupling is necessary tight between versions and artifacts.

Artifacts now know only the UUID of their datatypes, so there is no
string-sized dependent performance or attempt to maintain identical
static string pointers. The UUIDs are still generated deterministically
from the name via a UUIDv5 namespace.