Doc: include mermaid for graphs in rustdoc
Use a kludge to include the mermaid.js text-to-graph rendering library
in rustdocs both for normal `cargo doc` builds and for docs.rs builds.
While heraclitus is abandoned, this was done here a library with rather
complex docs structure as a test for use in future libraries.

Note that the test doc comment for this is on a private struct, so is
only visible when built with

    cargo doc --document-private-items